requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.22.0
sqlalchemy==1.4.54
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
bcrypt==3.2.2  # Pin for passlib compat
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.32.3
tenacity==8.2.3
pytz==2024.1
python-dateutil==2.8.2
redis==5.0.1
pytest==8.4.2
testcontainers[postgres]==4.13.2
apache-airflow==2.9.3
pydantic-settings